feat: move BASE_DIR to modules (makes more sense)
utils.py was removed, since it no longer contains something relevant to
the project.
