+++ modules/__init__.py
@@ -1,3 +1,7 @@
+from os.path import abspath, dirname, split
+
+BASE_DIR = split(dirname(abspath(__file__)))[0]
+
 BUTTON_TEXT = "resultados anteriores"
 DATASET_DIR = "datasets"
 DATASET_TRIS = f"{DATASET_DIR}/Tris.csv"
@@ -7,7 +11,6 @@ URL_TRIS = f"{URL_DOMAIN}/Tris/Resultados"
 URL_MELATE_RETRO = f"{URL_DOMAIN}/MelateRetro/Resultados"
 ID_TRIS = 60
 ID_MELATE_RETRO = 30
-
 TABLE_LOTTERY = "lottery"
 TABLE_DRAW = "draw"
