--- modules/utils.py
@@ -1,3 +1,0 @@
-from os.path import abspath, dirname, split
-
-BASE_DIR = split(dirname(abspath(__file__)))[0]
